test: move add_delete_account to conftest
